+++ src/FScript.CSharpInterop/LanguageServer/LspRuntimeExterns.fs
@@ -1,0 +1,8 @@
+namespace FScript.LanguageServer
+
+open FScript.Language
+open FScript.CSharpInterop
+
+module LspRuntimeExterns =
+    let forSourcePath (sourcePath: string) : ExternalFunction list =
+        InteropServices.runtimeExternsForSourcePath sourcePath