+++ src/FScript.LanguageServer/Program.cs
@@ -1,0 +1,4 @@
+using FScript.LanguageServer.CSharp;
+
+var server = new LspServer();
+server.Run();