+++ src/FScript.LanguageServer/Program.fs
@@ -1,6 +1,0 @@
-module FScript.LanguageServer.Program
-
-[<EntryPoint>]
-let main _ =
-    LspServer.run ()
-    0